add authorize

+++ post/views.py
@@ -1,17 +1,24 @@
 from django.shortcuts import get_object_or_404
 
-from rest_framework import viewsets
+from rest_framework import viewsets, status
 from rest_framework.response import Response
+from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 
 from .serializer import PostSerializer
 from .models import Post
 
 class PostViewset(viewsets.ViewSet):
+    authentication_classes = (JSONWebTokenAuthentication,)
     
     def create(self, request, id=None):
+        if request.user.is_anonymous:
+            return Response({'message': 'token is needed'}, status=status.HTTP_401_UNAUTHORIZED)
+        
+        print(request.user)
         title = request.POST.get('title')
         description = request.POST.get('description')
-        post = Post(title=title, description=description)
+        user = request.user
+        post = Post(title=title, description=description, user=user)
         post.save()
         
         serializer = PostSerializer(post)
@@ -25,6 +32,7 @@ class PostViewset(viewsets.ViewSet):
         return Response(serializer.data)
     
 class PostElementViewset(viewsets.ViewSet):
+    authentication_classes = (JSONWebTokenAuthentication,)
         
     def retrieve(self, request, id=None):
         post = get_object_or_404(Post, id=id)
@@ -51,11 +59,11 @@ class PostElementViewset(viewsets.ViewSet):
     def destroy(self, request, id=None):
         post = get_object_or_404(Post, id=id)
         
-        post.delete()
+        if post.user == request.user:
+            post.delete()
+            return Response({'message': 'ok'})
         
-        serializer = PostSerializer(post)
-        
-        return Response(serializer.data)
+        return Response({'message': 'token is needed'}, status=status.HTTP_401_UNAUTHORIZED)
